moved exporting to it's own component, broke it apart to make it easier to read
