+++ src/app/exporting/exporting.component.html
@@ -1,0 +1,5 @@
+<div class="export" [ngClass]="{'open':exportOpen}">
+	<button class="btn btn-primary" id="exportLink" (click)="exportData()">Export</button>
+	<div id="html-container" (click)="selectText('html-container')"></div>
+	<div id="json-container" (click)="selectText('json-container')"></div>
+</div>
