+++ src/app/exporting/exporting.component.less
@@ -1,0 +1,16 @@
+
+.export {
+  position: relative;
+  z-index: 1;
+}
+
+.export {
+  > div {
+    display: none;
+  }
+  &.open {
+    > div {
+      display: block;
+    }
+  }
+}
