+++ src/app/retro-board/retro-board.component.html
@@ -6,7 +6,7 @@
         <span>{{ bucket.name }}</span>
       </h4>
       <div class="list-group">
-        <div class="btn-groups" *ngFor="let note of bucket.notes | async">
+        <div class="note-wrapper" *ngFor="let note of bucket.notes | async">
           <button type="button" class="list-group-item" [class.list-group-item-success]="bucket.type === 'success'"
             [class.list-group-item-danger]="bucket.type === 'danger'" [class.list-group-item-info]="bucket.type === 'info'" (click)="openModal(edittemplate, bucket, note)">
             {{note.message}}
@@ -22,11 +22,7 @@
       </div>
     </div>
   </div>
-  <div class="export" [ngClass]="{'open':exportOpen}">
-    <button class="btn btn-primary" id="exportLink" (click)="exportData()">Export</button>
-    <div id="json-container" (click)="selectText('json-container')"></div>
-    <div id="html-container" (click)="selectText('html-container')"></div>
-  </div>
+  <exporting [jsonData]="jsonData"></exporting>
 </div>
 <ng-template #addtemplate>
   <div class="modal-header">
