+++ src/app/retro-board/retro-board.component.less
@@ -1,48 +1,13 @@
 @import '../../variables.less';
 
-.list-group-item {
-  margin-bottom: 15px;
-  border-radius: 10px;
-}
-
-.count {
-  display: inline;
-  float: right;
-  padding: 0 3px;
-}
-
-.list-group-item-new {
-  border: 2px dashed @border-color;
-  opacity: 0.5;
-  text-align: center;
-}
-
-.list-group button {
-  font-family: @font;
-}
-
-.export {
+.note-wrapper {
   position: relative;
-  z-index: 1;
-}
 
-.export {
-  > div {
-    display: none;
-  }
-  &.open {
-    > div {
-      display: block;
-    }
-  }
-}
-
-.btn-groups {
-  position: relative;
   button {
     position: absolute;
     right: 11px;
     top: calc(~'50% - 11px');
+
     &:nth-of-type(1) {
       padding-right: 69px;
       position: relative;
@@ -50,18 +15,29 @@
       top: initial;
       right: initial;
     }
-    &.good {
-      background: @green-med;
-      border: @green-dark;
-    }
-    &.bad {
-      background: @red-med;
-      border: @red-dark;
+  }
+
+  .list-group-item {
+    margin-bottom: 15px;
+    border-radius: 10px;
+    font-family: @font;
+    
+    .count {
+      display: inline;
+      float: right;
+      padding: 0 3px;
     }
   }
+
   strong {
     top: calc(~'50% - 10px');
     position: absolute;
     right: 40px;
   }
+}
+
+.list-group-item-new {
+  border: 2px dashed @border-color;
+  opacity: 0.5;
+  text-align: center;
 }
